third_party/WebKit/LayoutTests/accessibility/aria-flowto.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body id="body">

<h1 id="item1" aria-flowto="item2 item3"></h1>

<div id="item2">Content 2</div>
<div id="item3">Content 3</div>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that aria-flowto identifies the correct element");

    if (window.accessibilityController) {
        var item1 = accessibilityController.accessibleElementById("item1");
        var item2 = accessibilityController.accessibleElementById("item2");
        var item3 = accessibilityController.accessibleElementById("item3");

        shouldBeTrue("item1.ariaFlowToElementAtIndex(0).isEqual(item2)");
        shouldBeTrue("item1.ariaFlowToElementAtIndex(1).isEqual(item3)");
    }

</script>

</body>
</html>
